Put config tests in all_providers
That will allow adding more all_providers tests easily.
