+++ e2e_tests/tests/all_providers/config/mod.rs
@@ -10,7 +10,7 @@ use std::process::Command;
 use std::thread;
 use std::time::Duration;
 
-const CONFIG_TOMLS_FOLDER: &str = "tests/config/tomls";
+const CONFIG_TOMLS_FOLDER: &str = "tests/all_providers/config/tomls";
 const SERVICE_CONFIG_PATH: &str = "provider_cfg/tmp_config.toml";
 
 fn set_config(filename: &str) {
