+++ e2e_tests/tests/mod.rs
@@ -31,5 +31,4 @@
 #![allow(clippy::multiple_crate_versions)]
 
 mod all_providers;
-mod config;
 mod per_provider;
